+++ dealScraper/dealScraper.py
@@ -1,6 +1,6 @@
 # Imports
 import sys
-import json
+from pathlib import Path
 
 # Import local files
 from GUI import *
@@ -9,30 +9,37 @@ from whatsapp import WhatsAppSender
 
 
 def main():
-    # Local Variables
+    # Local File Variables
     params_file = 'params.json'
     number_file = 'number.txt'
+    params_path = Path(params_file)
+    number_path = Path(number_file)
 
-    # Create new params file
-
-    try:
-
+    # Checks for params.json and number.txt, if non-existent, loads GUI
+    if params_path.is_file() and number_path.is_file():
+        print("'params.json' and 'number.txt' currently exist in this local directory. \n"
+              "If you wish to update these and re-run the GUI, please delete them both from this folder and re-run the "
+              "application.")
+    else:
+        # Create new params file
+        print("Let's create your preference files!")
         gui()
 
+    try:
+        # Load preferences from local file params.json
         with open(params_file) as f:
             params = json.load(f)
             f.close()
 
         # Webscraper Call
-        absoluteURL = formatRequest(params)
-        soup = getProducts(absoluteURL)
+        absolute_url = format_request(params)
+        soup = get_products(absolute_url)
         data = parser(soup)
-        # data = list(data)
 
         with open(number_file) as f:
             phone_num = "+" + f.read()
 
-        # WhatsApp Sender Call
+        # WhatsApp Sender Call / Message Formatting
         ws = WhatsAppSender(data, phone_num)
         message = ""
         for items in data:
@@ -41,15 +48,9 @@ def main():
                 message += str(item)
                 message += "\n"
             message += "\n\n"
-        print(message)
-        '''
-        for item in data:
-            message += str(item)
-            message += "\n"
-        '''
-        # message = ws.format_scraped_info(data)
         ws.send_to_whatsapp(message)
 
+    # Any unexpected error closes application
     except:
         print("DealScraper encountered an error. Closing...")
         sys.exit()